Day 3 - Problem 2: Looks like Aggregate was silently hiding an int overflow. Switching the input to a long takes care of this.

+++ AdventOfCode.Day3/Problem2.cs
@@ -17,7 +17,7 @@ namespace AdventOfCode.Day3
                 new Tuple<int, int>(1, 2),
             };
 
-            var totalTrees = new List<int>();
+            var totalTrees = new List<long>();
             foreach (var step in steps)
             {
                 var stepX = step.Item1;
